fixes #2652 refactor client simplepool to address some small issues

+++ client/src/main/java/org/apache/hc/core5/util/copied/Args.java
@@ -31,26 +31,54 @@ import org.apache.hc.core5.http.copied.EntityDetails;
 
 import java.util.Collection;
 
+/**
+ * Utility class for argument validation.
+ */
 public class Args {
 
+    /**
+     * Checks an expression and throws IllegalArgumentException if false.
+     * @param expression the expression to check
+     * @param message the exception message
+     * @throws IllegalArgumentException if expression is false
+     */
     public static void check(final boolean expression, final String message) {
         if (!expression) {
             throw new IllegalArgumentException(message);
         }
     }
 
+    /**
+     * Checks an expression and throws IllegalArgumentException if false.
+     * @param expression the expression to check
+     * @param message the exception message format
+     * @param args the format arguments
+     * @throws IllegalArgumentException if expression is false
+     */
     public static void check(final boolean expression, final String message, final Object... args) {
         if (!expression) {
             throw new IllegalArgumentException(String.format(message, args));
         }
     }
 
+    /**
+     * Checks an expression and throws IllegalArgumentException if false.
+     * @param expression the expression to check
+     * @param message the exception message format
+     * @param arg the format argument
+     * @throws IllegalArgumentException if expression is false
+     */
     public static void check(final boolean expression, final String message, final Object arg) {
         if (!expression) {
             throw new IllegalArgumentException(String.format(message, arg));
         }
     }
 
+    /**
+     * Checks the content length of an entity.
+     * @param entityDetails the entity details
+     * @return the content length
+     */
     public static long checkContentLength(final EntityDetails entityDetails) {
         // -1 is a special value
         // 0 is allowed as well
@@ -58,6 +86,15 @@ public class Args {
                         "HTTP entity too large to be buffered in memory)");
     }
 
+    /**
+     * Checks if a value is within a range.
+     * @param value the value to check
+     * @param lowInclusive the lower bound
+     * @param highInclusive the upper bound
+     * @param message the exception message
+     * @return the value if within range
+     * @throws IllegalArgumentException if out of range
+     */
     public static int checkRange(final int value, final int lowInclusive, final int highInclusive,
                     final String message) {
         if (value < lowInclusive || value > highInclusive) {
@@ -67,15 +104,32 @@ public class Args {
         return value;
     }
 
+    /**
+     * Checks if a value is within a range.
+     * @param value the value to check
+     * @param lowInclusive the lower bound
+     * @param highInclusive the upper bound
+     * @param message the exception message
+     * @return the value if within range
+     * @throws IllegalArgumentException if out of range
+     */
     public static long checkRange(final long value, final long lowInclusive, final long highInclusive,
                     final String message) {
         if (value < lowInclusive || value > highInclusive) {
             throw illegalArgumentException("%s: %,d is out of range [%,d, %,d]", message, Long.valueOf(value),
-                            Long.valueOf(lowInclusive), Long.valueOf(highInclusive));
+                            Integer.valueOf((int) lowInclusive), Integer.valueOf((int) highInclusive));
         }
         return value;
     }
 
+    /**
+     * Checks if a character sequence contains no blanks.
+     * @param <T> the type of the character sequence
+     * @param argument the character sequence to check
+     * @param name the name of the argument
+     * @return the character sequence if no blanks
+     * @throws IllegalArgumentException if contains blanks
+     */
     public static <T extends CharSequence> T containsNoBlanks(final T argument, final String name) {
         if (argument == null) {
             throw illegalArgumentExceptionNotNull(name);
@@ -101,6 +155,14 @@ public class Args {
         return new IllegalArgumentException(name + " must not be null");
     }
 
+    /**
+     * Checks if a character sequence is not blank.
+     * @param <T> the type of the character sequence
+     * @param argument the character sequence to check
+     * @param name the name of the argument
+     * @return the character sequence if not blank
+     * @throws IllegalArgumentException if blank
+     */
     public static <T extends CharSequence> T notBlank(final T argument, final String name) {
         if (argument == null) {
             throw illegalArgumentExceptionNotNull(name);
@@ -111,6 +173,14 @@ public class Args {
         return argument;
     }
 
+    /**
+     * Checks if a character sequence is not empty.
+     * @param <T> the type of the character sequence
+     * @param argument the character sequence to check
+     * @param name the name of the argument
+     * @return the character sequence if not empty
+     * @throws IllegalArgumentException if empty
+     */
     public static <T extends CharSequence> T notEmpty(final T argument, final String name) {
         if (argument == null) {
             throw illegalArgumentExceptionNotNull(name);
@@ -121,6 +191,15 @@ public class Args {
         return argument;
     }
 
+    /**
+     * Checks if a collection is not empty.
+     * @param <E> the type of the collection elements
+     * @param <T> the type of the collection
+     * @param argument the collection to check
+     * @param name the name of the argument
+     * @return the collection if not empty
+     * @throws IllegalArgumentException if empty
+     */
     public static <E, T extends Collection<E>> T notEmpty(final T argument, final String name) {
         if (argument == null) {
             throw illegalArgumentExceptionNotNull(name);
@@ -131,6 +210,13 @@ public class Args {
         return argument;
     }
 
+    /**
+     * Checks if an integer is not negative.
+     * @param n the integer to check
+     * @param name the name of the argument
+     * @return the integer if not negative
+     * @throws IllegalArgumentException if negative
+     */
     public static int notNegative(final int n, final String name) {
         if (n < 0) {
             throw illegalArgumentException("%s must not be negative: %,d", name, n);
@@ -138,6 +224,13 @@ public class Args {
         return n;
     }
 
+    /**
+     * Checks if a long is not negative.
+     * @param n the long to check
+     * @param name the name of the argument
+     * @return the long if not negative
+     * @throws IllegalArgumentException if negative
+     */
     public static long notNegative(final long n, final String name) {
         if (n < 0) {
             throw illegalArgumentException("%s must not be negative: %,d", name, n);
@@ -145,6 +238,14 @@ public class Args {
         return n;
     }
 
+    /**
+     * Checks if an object is not null.
+     * @param <T> the type of the object
+     * @param argument the object to check
+     * @param name the name of the argument
+     * @return the object if not null
+     * @throws IllegalArgumentException if null
+     */
     public static <T> T notNull(final T argument, final String name) {
         if (argument == null) {
             throw illegalArgumentExceptionNotNull(name);
@@ -152,6 +253,13 @@ public class Args {
         return argument;
     }
 
+    /**
+     * Checks if an integer is positive.
+     * @param n the integer to check
+     * @param name the name of the argument
+     * @return the integer if positive
+     * @throws IllegalArgumentException if not positive
+     */
     public static int positive(final int n, final String name) {
         if (n <= 0) {
             throw illegalArgumentException("%s must not be negative or zero: %,d", name, n);
@@ -159,6 +267,13 @@ public class Args {
         return n;
     }
 
+    /**
+     * Checks if a long is positive.
+     * @param n the long to check
+     * @param name the name of the argument
+     * @return the long if positive
+     * @throws IllegalArgumentException if not positive
+     */
     public static long positive(final long n, final String name) {
         if (n <= 0) {
             throw illegalArgumentException("%s must not be negative or zero: %,d", name, n);
